[cleanup] Moved cost-base and cost-sum files to core folder
