+++ benchmark/bipedal-timings.cpp
@@ -21,7 +21,7 @@
 
 #include "crocoddyl/core/mathbase.hpp"
 
-#include "crocoddyl/multibody/costs/cost-sum.hpp"
+#include "crocoddyl/core/costs/cost-sum.hpp"
 #include "crocoddyl/multibody/costs/frame-placement.hpp"
 #include "crocoddyl/multibody/costs/state.hpp"
 #include "crocoddyl/multibody/costs/control.hpp"
