+++ bindings/python/crocoddyl/core/cost-base.hpp
@@ -6,10 +6,10 @@
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
 
-#ifndef BINDINGS_PYTHON_CROCODDYL_MULTIBODY_COST_BASE_HPP_
-#define BINDINGS_PYTHON_CROCODDYL_MULTIBODY_COST_BASE_HPP_
+#ifndef BINDINGS_PYTHON_CROCODDYL_CORE_COST_BASE_HPP_
+#define BINDINGS_PYTHON_CROCODDYL_CORE_COST_BASE_HPP_
 
-#include "crocoddyl/multibody/cost-base.hpp"
+#include "crocoddyl/core/cost-base.hpp"
 #include "crocoddyl/core/utils/exception.hpp"
 
 namespace crocoddyl {
