+++ bindings/python/crocoddyl/multibody/multibody.hpp
@@ -22,10 +22,8 @@ void exposeStateMultibody();
 void exposeActuationFloatingBase();
 void exposeActuationFull();
 void exposeActuationModelMultiCopterBase();
-void exposeCostAbstract();
 void exposeContactAbstract();
 void exposeImpulseAbstract();
-void exposeCostSum();
 void exposeContactMultiple();
 void exposeImpulseMultiple();
 void exposeDataCollectorMultibody();
