+++ include/crocoddyl/multibody/actions/free-fwddyn.hpp
@@ -17,10 +17,10 @@
 
 #include "crocoddyl/multibody/fwd.hpp"
 #include "crocoddyl/core/diff-action-base.hpp"
+#include "crocoddyl/core/costs/cost-sum.hpp"
 #include "crocoddyl/core/actuation-base.hpp"
 #include "crocoddyl/multibody/data/multibody.hpp"
 #include "crocoddyl/multibody/states/multibody.hpp"
-#include "crocoddyl/multibody/costs/cost-sum.hpp"
 #include "crocoddyl/core/utils/exception.hpp"
 
 namespace crocoddyl {
