+++ include/crocoddyl/multibody/costs/frame-translation.hpp
@@ -11,7 +11,7 @@
 #define CROCODDYL_MULTIBODY_COSTS_FRAME_TRANSLATION_HPP_
 
 #include "crocoddyl/multibody/fwd.hpp"
-#include "crocoddyl/multibody/cost-base.hpp"
+#include "crocoddyl/core/cost-base.hpp"
 #include "crocoddyl/multibody/states/multibody.hpp"
 #include "crocoddyl/multibody/data/multibody.hpp"
 #include "crocoddyl/multibody/frames.hpp"
